--- src/main/java/cj/task/sleact/core/workspace/controller/dto/request/CreateChannelReq.java
@@ -1,24 +1,0 @@
-package cj.task.sleact.core.workspace.controller.dto.request;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@NoArgsConstructor
-public class CreateChannelReq {
-
-    @NotBlank
-    String name;
-
-    @Builder
-    public CreateChannelReq(String name) {
-        this.name = name;
-    }
-
-}
