+++ src/test/java/cj/task/sleact/core/workspace/component/ChatComponentTest.java
@@ -1,5 +1,6 @@
 package cj.task.sleact.core.workspace.component;
 
+import cj.task.sleact.core.chat.component.ChatComponent;
 import cj.task.sleact.entity.ChannelChat;
 import cj.task.sleact.repository.ChannelChatRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -14,8 +15,11 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.then;
+import static org.mockito.Mockito.times;
 
 @ExtendWith(MockitoExtension.class)
 class ChatComponentTest {
@@ -53,12 +57,12 @@ class ChatComponentTest {
             given(channelChatRepository.findPagingAndFetchMemberByChannelId(anyLong())).willReturn(mockChats);
 
             // when
-//            List<ChatInfoRes> result = chatComponent.findPagingList(channelId, perPage, page);
+            List<ChannelChat> result = chatComponent.findPagingList(channelId, perPage, page);
 
             // then
-//            then(channelChatRepository).should(times(1)).findPagingAndFetchMemberByChannelId(channelId);
+            then(channelChatRepository).should(times(1)).findPagingAndFetchMemberByChannelId(channelId);
 
-//            assertThat(result).extracting(ChatInfoRes::getId).containsExactly(3L, 1L);
+            assertThat(result).extracting(ChannelChat::getId).containsExactly(3L, 1L);
         }
 
     }
