+++ src/test/java/cj/task/sleact/core/workspace/service/ChatServiceTest.java
@@ -1,10 +1,12 @@
 package cj.task.sleact.core.workspace.service;
 
-import cj.task.sleact.core.workspace.component.ChannelComponent;
-import cj.task.sleact.core.workspace.component.ChatComponent;
+import cj.task.sleact.core.channel.component.ChannelComponent;
+import cj.task.sleact.core.chat.component.ChatComponent;
+import cj.task.sleact.core.chat.controller.response.ChatInfoRes;
+import cj.task.sleact.core.chat.service.ChatService;
 import cj.task.sleact.core.workspace.component.WorkspaceComponent;
-import cj.task.sleact.core.workspace.controller.dto.response.ChatInfoRes;
 import cj.task.sleact.entity.Channel;
+import cj.task.sleact.entity.ChannelChat;
 import cj.task.sleact.entity.Workspace;
 import cj.task.sleact.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -18,8 +20,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.then;
+import static org.mockito.Mockito.times;
 
 @ExtendWith(MockitoExtension.class)
 class ChatServiceTest {
@@ -60,20 +66,20 @@ class ChatServiceTest {
             given(workspaceComponent.findWorkspaceByUrl(anyString())).willReturn(mockWorkspace);
             given(channelComponent.findByWorkspaceUrlAndChannelName(anyString(), anyString())).willReturn(mockChannel);
 
-            ChatInfoRes chat1 = Mockito.mock(ChatInfoRes.class);
-            ChatInfoRes chat2 = Mockito.mock(ChatInfoRes.class);
-            List<ChatInfoRes> mockChats = List.of(chat1, chat2);
-//            given(chatComponent.findPagingList(anyLong(), anyLong(), anyLong())).willReturn(mockChats);
+            ChannelChat chat1 = Mockito.mock(ChannelChat.class);
+            ChannelChat chat2 = Mockito.mock(ChannelChat.class);
+            List<ChannelChat> mockChats = List.of(chat1, chat2);
+            given(chatComponent.findPagingList(anyLong(), anyLong(), anyLong())).willReturn(mockChats);
 
             // when
-//            List<ChatInfoRes> result = chatService.findPagingList(workspaceUrl, channelName, perPage, page);
+            List<ChatInfoRes> result = chatService.findPagingList(workspaceUrl, channelName, perPage, page);
 
             // then
-//            then(workspaceComponent).should(times(1)).findWorkspaceByUrl(workspaceUrl);
-//            then(channelComponent).should(times(1)).findByWorkspaceUrlAndChannelName(workspaceUrl, channelName);
-//            then(chatComponent).should(times(1)).findPagingList(channelId, perPage, page);
-//
-//            assertThat(result).containsExactly(chat1, chat2);
+            then(workspaceComponent).should(times(1)).findWorkspaceByUrl(workspaceUrl);
+            then(channelComponent).should(times(1)).findByWorkspaceUrlAndChannelName(workspaceUrl, channelName);
+            then(chatComponent).should(times(1)).findPagingList(channelId, perPage, page);
+
+            assertThat(result).containsExactly(chat1, chat2);
         }
 
     }
